+++ ChangeLog.md
@@ -7,5 +7,6 @@ All notable changes to this project will be documented in this file.
 
 ## Version 1.4
 
+- FIX : Line break in delivery address when several lines in public note *26-08-2021* - 1.4.1
 - FIX - Compatibility V14 : Edit the descriptor: editor_url and family - *2021-06-10* - 1.4.1
 - FIX : missing GETPOST check parameters (none)